feat: pass test for new standard DOB/0
